pyproject.toml
# https://packaging.python.org/en/latest/guides/writing-pyproject-toml/
[project]
name = "cqrs-test"
version = "0.1.0.dev0"
requires-python = ">= 3.12"
dependencies = [
  # backend
  "granian[reload]==1.3.1",
  "django-ninja==1.1.0",
]
optional-dependencies.dev = [
  "daphne==4.1.2",
]